U-RWKV/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
einops>=0.3.0
timm>=0.4.12
wandb>=0.12.0
numpy>=1.19.2
pillow>=8.0.0
tqdm>=4.50.0
scikit-learn>=0.24.0
matplotlib>=3.3.0 